Breaking change: Structure remodel
