--- API.sln
@@ -1,43 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32228.430
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "API", "API\API.csproj", "{9DB783F2-90F1-4D78-8F9A-741AC4243A75}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application", "Application\Application.csproj", "{0D2A020E-B5EF-49E9-9E85-E4611E368A52}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "Domain\Domain.csproj", "{A774B108-866C-45CA-AD81-46BDAEDA9CF9}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Infrastructure", "Infrastructure\Infrastructure.csproj", "{4FA29B62-584A-4BDC-AB32-283C767DDE54}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9DB783F2-90F1-4D78-8F9A-741AC4243A75}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9DB783F2-90F1-4D78-8F9A-741AC4243A75}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9DB783F2-90F1-4D78-8F9A-741AC4243A75}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9DB783F2-90F1-4D78-8F9A-741AC4243A75}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0D2A020E-B5EF-49E9-9E85-E4611E368A52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0D2A020E-B5EF-49E9-9E85-E4611E368A52}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0D2A020E-B5EF-49E9-9E85-E4611E368A52}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0D2A020E-B5EF-49E9-9E85-E4611E368A52}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A774B108-866C-45CA-AD81-46BDAEDA9CF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A774B108-866C-45CA-AD81-46BDAEDA9CF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A774B108-866C-45CA-AD81-46BDAEDA9CF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A774B108-866C-45CA-AD81-46BDAEDA9CF9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4FA29B62-584A-4BDC-AB32-283C767DDE54}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4FA29B62-584A-4BDC-AB32-283C767DDE54}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4FA29B62-584A-4BDC-AB32-283C767DDE54}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4FA29B62-584A-4BDC-AB32-283C767DDE54}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {63F4B154-1C79-487B-93CD-123DD873FA58}
-	EndGlobalSection
-EndGlobal
