--- API/Settings/MongoDbConfig.cs
@@ -1,8 +1,0 @@
-﻿namespace API.Settings
-{
-    public class MongoDbConfig
-    {
-        public string Name { get; init; }
-        public string ConnectionString { get; init; }
-    }
-}
