--- API/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "MongoDbConfig": {
-    "Name": "Exadel-Budget",
-    "ConnectionString": "mongodb+srv://exadel:exadel@budgetcluster.bjelq.mongodb.net/?retryWrites=true&w=majority"
-  }
-}
