--- Domain/Enums/CategoryTypes.cs
@@ -1,8 +1,0 @@
-﻿namespace Domain.Enums
-{
-    public enum CategoryTypes
-    {
-        Credit,
-        Debit
-    }
-}
