--- Domain/Enums/LimitPeriods.cs
@@ -1,9 +1,0 @@
-﻿namespace Domain.Enums
-{
-    public enum LimitPeriods
-    {
-        Daily,
-        Weekly,
-        Monthly
-    }
-}
