--- Domain/Enums/NotificationTypes.cs
@@ -1,10 +1,0 @@
-﻿namespace Domain.Enums
-{
-    public enum NotificationTypes
-    {
-        LimitExceeded,
-        MonthlyExpense,
-        Encouraging,
-        Other
-    }
-}
