--- Domain/Models/Notification.cs
@@ -1,14 +1,0 @@
-using Domain.Enums;
-using MongoDB.Bson.Serialization.Attributes;
-using System.ComponentModel.DataAnnotations;
-
-namespace Domain.Models
-{
-    public class Notification
-    {
-        [BsonId]
-        public Guid Id { get; set; }
-        [Required]
-        public NotificationTypes NotificationType { get; set; }
-    }
-}
