--- Infrastructure/Models/ApplicationUser.cs
@@ -1,9 +1,0 @@
-﻿using AspNetCore.Identity.MongoDbCore.Models;
-
-namespace Infrastructure.Models
-{
-    public class ApplicationUser : MongoIdentityUser<Guid>
-    {
-        public Guid UserId { get; set; }
-    }
-}
