+++ src/BudgetManager.sln
@@ -1,0 +1,49 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32602.215
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetManager.API", "BudgetManager\BudgetManager.API\BudgetManager.API.csproj", "{CD215130-F3C6-499F-BFAF-1C631B46729C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetManager.CQRS", "BudgetManager\BudgetManager.CQRS\BudgetManager.CQRS.csproj", "{F7260E6F-0D7C-4BB9-9C15-2CA1D0164630}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetManager.DataAccess", "BudgetManager\BudgetManager.DataAccess\BudgetManager.DataAccess.csproj", "{EFB156F4-CD75-4ADE-8193-893FA583E1CD}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetManager.Model", "BudgetManager\BudgetManager.Model\BudgetManager.Model.csproj", "{109C6D98-F25E-4D0B-8590-D0AB1245E06D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BudgetManager.SDK", "BudgetManager\BudgetManager.SDK\BudgetManager.SDK.csproj", "{69148D4F-FF17-4468-B5EB-EA8910E83BC9}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CD215130-F3C6-499F-BFAF-1C631B46729C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CD215130-F3C6-499F-BFAF-1C631B46729C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CD215130-F3C6-499F-BFAF-1C631B46729C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CD215130-F3C6-499F-BFAF-1C631B46729C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F7260E6F-0D7C-4BB9-9C15-2CA1D0164630}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F7260E6F-0D7C-4BB9-9C15-2CA1D0164630}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F7260E6F-0D7C-4BB9-9C15-2CA1D0164630}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F7260E6F-0D7C-4BB9-9C15-2CA1D0164630}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EFB156F4-CD75-4ADE-8193-893FA583E1CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EFB156F4-CD75-4ADE-8193-893FA583E1CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EFB156F4-CD75-4ADE-8193-893FA583E1CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EFB156F4-CD75-4ADE-8193-893FA583E1CD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{109C6D98-F25E-4D0B-8590-D0AB1245E06D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{109C6D98-F25E-4D0B-8590-D0AB1245E06D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{109C6D98-F25E-4D0B-8590-D0AB1245E06D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{109C6D98-F25E-4D0B-8590-D0AB1245E06D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{69148D4F-FF17-4468-B5EB-EA8910E83BC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{69148D4F-FF17-4468-B5EB-EA8910E83BC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{69148D4F-FF17-4468-B5EB-EA8910E83BC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{69148D4F-FF17-4468-B5EB-EA8910E83BC9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {218A3F9E-DEB4-4B3F-9C48-A638693F1827}
+	EndGlobalSection
+EndGlobal
